requirements.txt
# Gemini-Drive MCP 프로젝트 필수 라이브러리
# 설치 방법: pip install -r requirements.txt

# MCP (Model Context Protocol) SDK
# Claude Desktop과의 통신을 위한 프로토콜
mcp>=0.1.0

# Google AI (Gemini) SDK - Gemini API를 사용하기 위한 라이브러리
google-generativeai>=0.3.0

# Google 인증 관련 라이브러리들
# Google API를 사용하기 위한 인증 처리
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Google Drive API - Drive에 파일을 읽고 쓰기 위한 라이브러리
google-api-python-client>=2.100.0

# 환경 변수 관리 - .env 파일에서 API 키를 안전하게 불러오기
python-dotenv>=1.0.0

# HTTP 요청 라이브러리 - API 호출에 사용
requests>=2.31.0

# 터미널 색상 출력 - 사용자에게 보기 좋은 메시지 표시 (레거시 코드 호환성)
colorama>=0.4.6

# 비동기 프로그래밍 지원
asyncio>=3.4.3

# 테스트 라이브러리 (개발용)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Flask 웹 프레임워크 - 웹 UI를 위한 라이브러리
Flask>=3.0.0
Flask-CORS>=4.0.0
